# APISistemaVenta/SistemaVenta.API/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SistemaVenta.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Activo/Guardar": {
      "post": {
        "tags": [
          "Activo"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ActivoDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Activo/Editar": {
      "put": {
        "tags": [
          "Activo"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ActivoDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Activo/Filtros": {
      "get": {
        "tags": [
          "Activo"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "buscarPor",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "encargado",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "cliente",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Activo/EditarEstado": {
      "put": {
        "tags": [
          "Activo"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ActivoDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Activo/Eliminar/{id}": {
      "delete": {
        "tags": [
          "Activo"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Cliente/Guardar": {
      "post": {
        "tags": [
          "Cliente"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ClienteDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cliente/Editar": {
      "put": {
        "tags": [
          "Cliente"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ClienteDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cliente/Eliminar/{id}": {
      "delete": {
        "tags": [
          "Cliente"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DashBoard/Resumen": {
      "get": {
        "tags": [
          "DashBoard"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "interv",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DashBoard/ResumenMes": {
      "get": {
        "tags": [
          "DashBoard"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "mes",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "año",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DashBoard/ResumenAnual": {
      "get": {
        "tags": [
          "DashBoard"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "año",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/DashBoard/ResumenVendedores": {
      "get": {
        "tags": [
          "DashBoard"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "año",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Producto/Guardar": {
      "post": {
        "tags": [
          "Producto"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProductoDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Producto/Editar": {
      "put": {
        "tags": [
          "Producto"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProductoDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Producto/Eliminar/{id}": {
      "delete": {
        "tags": [
          "Producto"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "ActivoDTO": {
      "type": "object",
      "properties": {
        "idActivo": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "marca": {
          "type": "string"
        },
        "codigo": {
          "type": "string"
        },
        "cliente": {
          "type": "string"
        },
        "encargado": {
          "type": "string"
        },
        "esContrato": {
          "format": "int32",
          "type": "integer"
        },
        "esActivo": {
          "format": "int32",
          "type": "integer"
        },
        "fechaCaducidad": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "ClienteDTO": {
      "type": "object",
      "properties": {
        "idCliente": {
          "format": "int32",
          "type": "integer"
        },
        "nombreCompleto": {
          "type": "string"
        },
        "correo": {
          "type": "string"
        },
        "contacto": {
          "type": "string"
        },
        "esActivo": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
    "ProductoDTO": {
      "type": "object",
      "properties": {
        "idProducto": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "idCategoria": {
          "format": "int32",
          "type": "integer"
        },
        "compatibilidad": {
          "type": "string"
        },
        "descripcionCategoria": {
          "type": "string"
        },
        "stock": {
          "format": "int32",
          "type": "integer"
        },
        "precio": {
          "type": "string"
        },
        "esActivo": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
  }
}